+++ src/language/dateformat.ts
@@ -1,0 +1,25 @@
+export const dayNames = (lang: string) => {
+    if(lang == 'es-ES') return [
+        'Dom', 'Lun', 'Mar', 'Mié', 'Jue', 'Vie', 'Sáb',
+        'Domingo', 'Lunes', 'Martes', 'Miércoles', 'Jueves', 'Viernes', 'Sábado'
+    ]
+    return [
+        "Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat",
+        "Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"
+    ]
+}
+
+export const monthNames = (lang: string) => {
+    if(lang == 'es-ES') return [ 
+        'Ene', 'Feb', 'Mar', 'Abr', 'May', 'Jun', 'Jul', 'Ago', 'Sep', 'Oct', 'Nov', 'Dic',
+        'Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio', 'Julio', 'Agosto', 'Septiembre', 'Octubre', 'Noviembre', 'Diciembre'
+    ]
+    return [ 
+        "Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec",
+        "January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"
+    ]
+}
+
+export const timeNames = () => {
+    return [ 'a', 'p', 'am', 'pm', 'A', 'P', 'AM', 'PM' ]
+}