+++ src/types/config.d.ts
@@ -1,0 +1,1 @@
+type Config = { language: Lang }