+++ src/types/language.d.ts
@@ -1,0 +1,1 @@
+type Lang = 'es-ES' | 'en-EN'