+++ src/utils/generators/generateNumberBetween.ts
@@ -2,7 +2,7 @@
  * Generate number between two numbers
  *
  * @param {number} min `number` Minimum number
- * @param {number} min `number` Máximum number
+ * @param {number} min `number` Maximum number
  * @return {number} number `number` Number between min and max
  * @example
  * OU.generateNumberBetween(300, 500) // Return 382