+++ CHANGELOG.md
@@ -1,18 +1,50 @@
 # Changelog
 
+### 0.3.0
+
+Release 25 sep 2023
+
+  * Add [`containDigits`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#containDigits) to `validators` related utils
+  * Add [`containNormalcaseLetters`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#containNormalcaseLetters) to `validators` related utils
+  * Add [`containSpecialChars`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#containSpecialChars) to `validators` related utils
+  * Add [`containUppercaseLetters`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#containUppercaseLetters) to `validators` related utils
+  * Add [`isEmailValid`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#isEmailValid) to `validators` related utils
+  * Add [`isPasswordValid`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#isPasswordValid) to `validators` related utils
+  * Add [`isUsernameValid`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#isUsernameValid) to `validators` related utils
+  
+  * Add [`extractBase64FileData`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#extractBase64FileData) to `data` related utils
+  * Add [`formatBytes`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#formatBytes) to `data` related utils
+  * Add [`removeHTML`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#removeHTML) to `data` related utils
+
+  * Add [`getPasswordStrength`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#getPasswordStrength) to `global` related utils
+
+  * Add [`calculateReadingTime`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#calculateReadingTime) to `time` related utils
+  * Add [`formatActualDate`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#formatActualDate) to `time` related utils
+  * Add [`formatTimestamp`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#formatTimestamp) to `time` related utils
+  * Add [`getRelativeTime`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#getRelativeTime) to `time` related utils
+  * Add [`getTimestamp`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#getTimestamp) to `time` related utils
+  * Add [`sleep`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#sleep) to `time` related utils
+
+  * Rename `extractBase64Data` to [`convertBase64Data`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#convertBase64Data) to `converters` related utils
+
+### 0.2.2
+
+Release 23 sep 2023
+  * Remove github actions to simplify the process
+
 ### 0.2.0
 
 Release 23 sep 2023
-  * Add [`generateCode`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#generateCode) to generators related utils
-  * Add [`generateId`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#generateId) to generators related utils
-  * Add [`generateNumberBetween`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#generateNumberBetween) to generators related utils
-  * Add [`convertTextToSlug`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#convertTextToSlug) to converters related utils
+  * Add [`generateCode`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#generateCode) to `generators` related utils
+  * Add [`generateId`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#generateId) to `generators` related utils
+  * Add [`generateNumberBetween`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#generateNumberBetween) to `generators` related utils
+  * Add [`convertTextToSlug`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#convertTextToSlug) to `converters` related utils
 
 ### 0.1.3
 
 Release 23 sep 2023
-  * Add [`convertTimestampToDate`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#convertTimestampToDate) to time related utils
-  * Add [`extractBase64Data`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#extractBase64Data) to data related utils
+  * Add [`convertTimestampToDate`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#convertTimestampToDate) to `time` related utils
+  * Add [`extractBase64Data`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#extractBase64Data) to `data` related utils
 
 ### 0.1.2